+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 - Added `url` method to `Comment` model
+- Added `srcset` method to `Upload` model
 
 ## [2.1.1] - 2015-10-27
 ### Fixed